Moved all existing tests into test/unit
