+++ test/unit/services/proxy/processors/parse-user-agent.test.ts
@@ -1,6 +1,6 @@
 import {describe, it, expect} from 'vitest';
-import * as processors from '../../../../src/services/proxy/processors/parse-user-agent';
-import {FastifyRequest, HttpProxyRequest} from '../../../../src/types';
+import * as processors from '../../../../../src/services/proxy/processors/parse-user-agent';
+import {FastifyRequest, HttpProxyRequest} from '../../../../../src/types';
 
 const {parseUserAgent} = processors;
